requirements.txt
asgiref==3.6.0
Django==4.1.7
django-crispy-forms==2.0
django-filter==23.2
greenlet==2.0.2
numpy==1.24.3
pandas==2.0.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
sqlparse==0.4.3
typing_extensions==4.6.3
tzdata==2023.3
